feat!: support lazy-loading plugin configs
By:

  - Not loading `opt` plugins by default.
  - Providing a `configure(name)` function.
